Correciones en el fichero de datos iniciales
 - Correciones en el foramto YAML del fichero datos-iniciales.yaml
 - Actualizada la version de JBehave a 3.8

+++ app/models/Usuario.java
@@ -43,8 +43,8 @@ public class Usuario extends Model {
 
 	@Required
 	@Column(nullable = false, name = "num_tarjeta")
-	public Integer numTarjeta;
-	
+	public Long numTarjeta;
+
 	@OneToMany(mappedBy = "usuario")
 	public List<Entrada> entradas = new ArrayList<Entrada>();
